Restructuring, added cups string output Todo, incorporate new common cups strings Todo, move ports to constants
